% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{SURVFIT}
\alias{SURVFIT}
\title{Doubly Sparse Survival Rule Extraction}
\usage{
SURVFIT(
  formula = formula,
  data = data,
  rulelength = 3,
  doubly.sparse = FALSE,
  gamma = NULL,
  lambda1 = NULL,
  lambda2 = NULL,
  crossvalidate = TRUE,
  nfolds = 5,
  num_toprules = 16,
  num_totalrules = 2000,
  input_rule_list = FALSE,
  rule_list = NULL,
  ntree = 200,
  digit = 10,
  seed = NULL,
  nodesize = NULL,
  ...
)
}
\arguments{
\item{formula}{\code{formula}. The model formula specifying time, status and
dependent variables of the form \code{Surv(time, status)~ x1 + x2 + .. }}

\item{data}{\code{data.frame}. Training data.}

\item{rulelength}{\code{Integer}. Maximum length of the rule. (Default = 3)}

\item{doubly.sparse}{\code{Logical} for whether double sparsity required.
(Default = FALSE)}

\item{gamma}{\code{Numeric} or \code{list}. Hyperparameter (Default = NULL)}

\item{lambda1}{\code{Numeric} or \code{list}. Hyperparameter (Default = NULL)}

\item{lambda2}{\code{Numeric} or \code{list}. Hyperparameter (Default =
NULL)}

\item{crossvalidate}{\code{Logical}. Whether crossvalidation to be done to
find hyperparameters. (Default = TRUE)}

\item{nfolds}{\code{Integer}. Number of cross validation folds. (Default = 5)}

\item{num_toprules}{\code{Integer}. Number of rules extracted. (Default = 16)}

\item{num_totalrules}{\code{Integer}. Number of rules considered. (Default =
2000)}

\item{input_rule_list}{\code{Logical} Whether rule list supplied. (Default =
FALSE)}

\item{rule_list}{\code{List}. List of supplied rules. (Default = NULL)}

\item{ntree}{\code{Integer} .Number of trees built}

\item{digit}{\code{Integer}. Decimal points.}

\item{seed}{\code{Numeric}. Seed for reproducible experiments.}

\item{nodesize}{\code{Integer}. (Default = NULL)}

\item{...}{Other inputs}
}
\value{
Object of class \code{list} with elements
  \item{\code{top_rules}}{List of top \code{\link{num_toprules}} rules}
  \item{\code{all_rules}}{List of all \code{\link{num_totalrules}} rules}
  \item{\code{beta}}{Coefficients of \link{all_rules} in the model}
}
\description{
\code{SURVFIT} extracts a doubly sparse (sparse in both number of rules and
in number of variables in the rules) survival rule ensemble from survival
data
}
\examples{
## For ovarian data from survival package.
SURVFIT(Surv(futime, fustat) ~ ., data = ovarian)

}
